requirements.txt
# requirements.txt
torch>=2.0.0
numpy>=1.21.0
pyyaml>=6.0
colorama>=0.4.4
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
pandas>=1.3.0
tensorboard>=2.13.0
opencv-python>=4.5.3